requirements.txt
fastapi
requests
python-dotenv
pyyaml
chromadb
ollama
pypdf
pandas